test: update aio payload size golden to reflect Angular update (#45461)
Updates the AIO payload size goldens to reflect the Angular update
affecting both the `aio` and `aio_local` build (due to CLI devkit
changes affecting both goldens, and the Angular update also affecting
the non-local `aio` build / job).

Overall, the Material themes got a little larger, but this is
not caused by the framwork, so not relevant here.

PR Close #45461